Reduce duplication between target-specific modules
